[fix] ファイル分割・フォルダ構成変更
