+++ src/libs/apiClient.ts
@@ -1,6 +1,6 @@
 import { createClient } from "microcms-js-sdk";
 
-export const client = createClient({
+export const apiClient = createClient({
   serviceDomain: process.env.MICROCMS_SERVICE_DOMAIN ?? "",
   apiKey: process.env.MICROCMS_API_KEY ?? "",
 });
